Fixed bugs
Fixed login/creation bugs:

When the input field was blank, the computer was processing too fast and using those blank fields as the username and password.  Also for error messages like "try again..." it would do that.  Now, build in defense by checking basic username and password characteristics (positive length, not the same, no spaces)

Fixed bug for playback slider:

Not working for mouse-click; after looking into the ofxDatGui class, realized that because I set the focus (in a different gui, though for some reason it still affected this one) on the login_box, it would not work; unset the focus from login_box after use was over and now works fine